Updated compare
