+++ pyprocar/procarplot/__init__.py
@@ -1,2 +1,2 @@
 from .procarplot import ProcarPlot
-from .procarplotcompare import ProcarPlotCompare
+
