+++ pyprocar/procarplotcompare/__init__.py
@@ -1,0 +1,1 @@
+from .procarplotcompare import ProcarPlotCompare
